fixed issues and added auth tests

+++ src/controllers/auth.controller.ts
@@ -13,9 +13,16 @@ const refreshTokenSecret = process.env.JWT_REFRESH_SECRET!;
  */
 export const login = async (req: Request, res: Response): Promise<any> => {
   const { email, password } = req.body;
-  console.log("ENV", accessTokenSecret, refreshTokenSecret);
   try {
+    if (!email) {
+      return res.status(400).json({ msg: "Email is required" });
+    }
+    if (!password) {
+      return res.status(400).json({ msg: "Password is required" });
+    }
+
     const user = await User.findOne({ email });
+
     if (!user) {
       return res.status(404).json({ msg: "User not found" });
     }
@@ -33,7 +40,7 @@ export const login = async (req: Request, res: Response): Promise<any> => {
     // Generate refresh token
     const refreshToken = jwt.sign({ userId: user._id }, refreshTokenSecret);
 
-    // Speichere den Refresh Token in der Datenbank
+    // Save the refresh token in the database
     const tokenEntry = new Token({ token: refreshToken, userId: user._id });
     await tokenEntry.save();
 
@@ -59,7 +66,7 @@ export const refreshToken = async (
   }
 
   try {
-    // Prüfe, ob der Token in der Datenbank existiert
+    // Check if the token exists in the database
     const storedToken = await Token.findOne({ token });
     if (!storedToken) {
       return res.status(403).json({ msg: "Invalid refresh token" });
@@ -85,7 +92,10 @@ export const logout = async (req: Request, res: Response): Promise<any> => {
   const { token } = req.body;
 
   try {
-    // Lösche den Token aus der Datenbank
+    if (!token) {
+      return res.status(400).json({ msg: "Token is required for logout" });
+    }
+    // Delete the token from the database
     await Token.findOneAndDelete({ token });
 
     res.json({ msg: "Logout successful" });
